requirements.txt
PyYAML==3.11
nose==1.3.7
pycrypto==2.6.1
wsgiref==0.1.2
